requirements.txt
requests>=2.25.0
Pillow>=8.0.0

